+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/routes/definitions/base/home.js
@@ -17,7 +17,7 @@
   home: {
     patterns: ['(?*)', 'workspaces(/)'],
     component:
-      'react-component/container/workspaces-container/workspaces-container',
+      'react-component/workspaces/container',
     menu: {
       component: 'component/workspaces-menu/workspaces-menu.view',
     },
