+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/workspaces/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './workspaces-container'
+export { default } from './container'
